+++ src/main/java/de/samply/share/client/quality/report/QualityReportSchedulerFormat.java
@@ -1,0 +1,7 @@
+package de.samply.share.client.quality.report;
+
+public enum QualityReportSchedulerFormat {
+  YEAR,
+  MONTH,
+  ALL
+}